+++ packages/core/config/services.php
@@ -22,6 +22,7 @@ return static function (ContainerConfigurator $container): void {
                 __DIR__ . '/../src/Entity/',
                 __DIR__ . '/../src/Module/{ModuleBlueprint,CollectionBlueprint,FieldBlueprint}.php',
                 __DIR__ . '/../src/Money/Amount.php',
+                __DIR__ . '/../src/Numbering/NumberFormat.php',
                 __DIR__ . '/../src/Record/{Record,Derivation}.php',
                 __DIR__ . '/../src/Validation/UniqueFieldValue.php',
                 __DIR__ . '/../src/XiviCoreBundle.php',